requirements.txt
bravado-core==4.7.3
functools32==3.2.3.post2
jsonschema==2.6.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
rfc3987==1.3.7
simplejson==3.11.1
six==1.10.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
webcolors==1.7
